pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auth_api_django"
version = "0.1.0"
description = "Authentication API using Django"
requires-python = ">=3.8"
dependencies = [
    "Django>=3.2",
    "djangorestframework>=3.12",
    "djangorestframework-simplejwt>=4.6",
    "django-cors-headers>=3.7",
    "django-allauth>=0.44",
    "django-login-history2>=1.0",
]

[tool.setuptools]
packages = ["auth_api_django"]
